pyproject.toml
[tool.poetry]
name = "skypro-python-lesson-27"
version = "0.1.0"
description = ""
authors = ["Oleg Khayatov <okhayatov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.2"
djangorestframework = "^3.13.1"
psycopg2 = "^2.9.3"
Pillow = "^9.0.1"
djangorestframework-simplejwt = "^5.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
